+++ elmyra.ip.access.epo/CHANGES.rst
@@ -10,6 +10,7 @@ development
     - /today, /week, /month, /year
     - /country/us
     - /cpc/"H04B7-15542"
+- ux: link to espacenet worldwide (biblio data)
 
 0.15.0
 ======